requirements.txt
asgiref==3.2.7
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.41.0
django-braces==1.14.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-markdown-deux==1.0.5
django-oauth-toolkit==1.3.2
django-rest-framework-social-oauth2==1.1.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.14.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
gunicorn==20.0.4
httplib2==0.17.3
idna==2.9
importlib-metadata==1.6.0
isodate==0.6.0
Markdown==3.2.2
markdown2==2.3.9
oauthlib==3.1.0
Pillow==7.1.2
protobuf==3.12.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2020.1
razorpay==1.2.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
schedule==0.6.0
shortid==0.1.2
shortuuid==1.0.1
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
zipp==3.1.0
